backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx>=0.27.0
sse-starlette>=2.0.0
sendgrid>=6.11.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
PyJWT>=2.8.0
bcrypt>=4.1.0
cryptography>=42.0.0
redis>=5.0.0
rq>=1.16.1
psycopg[binary]>=3.2.1
psycopg_pool>=3.2.1
